workflows: see if setting BUILDAH_FORMAT creates oci images on quay.io
At the moment the images in samba-server appear to have manifest format
"v2s1" which is rather suboptimal. We would prefer oci or at least
docker "v2s2" (version 2 schema 2).

This change is an experiment to see if tweaking this variable gives
us a better result when the CI job pushes to quay.io.

Signed-off-by: John Mulligan <jmulligan@redhat.com>