server: Update checkout hash for the sambacc repo
When building the server image, the git hash for the sambacc repo is
incorrectly set to an old commit. As a result, the server image on
quay.io/samba.org/samba-server:latest
contains an old outdated version of sambacc.

Use the latest commit hash instead when building the server images.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>