server: make Containerfile capable of building default or nightly
Incorporate the new shell scripts into the container build file and
use them, along with a few new build args to construct a container
image based on the given build parameters. These parameters mean
that the single Containerfile can build with our default set of
samba packages or the nightly builds.
Extending this is a matter of extending the shell scripts and
calling the build command with the proper build args.

Signed-off-by: John Mulligan <jmulligan@redhat.com>