server: remove older dockerfile and support files
The previous dockerfile based on centos and shell script was a fantastic
start to this project. However it hasn't been maintained and doesn't
have the capabilities of the sambacc based images that are needed for
features like AD support or CTDB. Since the files will live on in version
history its better to remove them from the tree as not to serve as an
attractive nuisance for future viewers.

If we ever need cross/different distro support for our base images
we should tackle it later, and base it on the current Containerfile.

Signed-off-by: John Mulligan <jmulligan@redhat.com>