server: enable install-sambacc.sh to make use of rpm
If the dist dir has a rpm for sambacc as opposed to or in addition to a
wheel file, prefer to install sambacc from the rpm because this
is an rpm based distro and its good to pull everything in the same way.

This fixes a couple of bugs while were at it.

Signed-off-by: John Mulligan <jmulligan@redhat.com>