toolbox: use fedora36 as base image
When using samba-client as base image for toolbox, we get an enigmatic
failure due to dnf dependencies conflict. This does not appear to be
related to samba packages but rather to something in dnf internal
database. Avoid this problem by using fedora:36 as base image.

Fixes also a minor type in container-image workflow.

Refs:
  https://github.com/samba-in-kubernetes/samba-container/issues/96

Signed-off-by: Shachar Sharon <ssharon@redhat.com>