makefile: attempt to make the build rules more regular
Update the Makefile to improve the regularity and DRY-ness of the
container build rules. Each build rule is now made up of common
top-level variables, of which mostly just the names and default build
args change.

I think this improves the readability of the file as well as making it
easier to include the Makefile and then re-use the _img_build target
rather than copy and paste the podman build + buildfile creation commands.

Signed-off-by: John Mulligan <jmulligan@redhat.com>