Makefile: generalise build mechanisms
This abstracts out a few settings  a bit.
it is intended to make future additions of other build flavors easier.

Adding a new build flavor (OS) will essentially amount to adding
OS-specific containerfiles and a os-specific Makefile that includes the
main Makefile and sets the following  two override variables:
* OS_NAME
* REPO_BASE

see Makefile.opensuse for an example.

Signed-off-by: Michael Adam <obnox@samba.org>