Makefile: improve buildfile handling, including cleanup
This change improves the definitions of buildfile variables by
introducing a variable for the filename prefix, '.build'.

As a side effect, the clean target which previously just deleted a
somewhat arbitrary explicit list of buildfiles and was therefore
error-prone, could be
made more systematic and robust by introducing the use of a wildcard.
This way, the clean target should continue working even if future
changes change the exact names of buildfiles.

Signed-off-by: Michael Adam <obnox@samba.org>