makefile: copy common files into context directory
Copy the `images/common` directory into `<context>/.common` before
building the image. This provides the image builds with the content of the
common dir at build time.

This patch was previously implemented using a 2nd context dir using the
multiple build context support present in recent versions of docker and
podman (see:
https://www.docker.com/blog/dockerfiles-now-support-multiple-build-contexts/
)
Unfortunately, the CI version of dockr (and podman) were not recent
enough to have this feature. This left me with a few options:
* Update the CI versions and force all contributors to use newer
  versions
* Rework all the paths in container files to be relative to images
  rather than `images/<tgt>`
* Make a little build time hack to make common files available to the
  existing context dir

The first option would have been annoying to implement and potentially
annoying to contributors. The second option was feasible but seems more
disruptive then needed. The final option is a bit hacky but also serves
as a small breadcrumb to come back and use a build context again some
time in the future.

Signed-off-by: John Mulligan <jmulligan@redhat.com>