makefile: add ability to build arm64 (on certain conditions)
Add a standard way to request arm64 (or other) architectures using
the `--arch` flag to podman. This makefile change does *nothing* to
attempt to make that flag work on a particular system. If the flag
works then this makefile will build the image and save it, with
the correct arch recorded in the recently added "fully qualified
image name".

This change also does nothing with with image, no tests, no creating
of manifest lists to create a something pushable, nothing.
I tried to start some of that, which lead to me attempting to lightly
reorganize the makefile, which lead me to attempting to heavily
reorganize the makefile, and that was a huge mess and a terrible slog
and the delta would have been impossible to follow. Thus I'm now
creating the smallest delta I can think of that still gets the work
to support arm64 builds started.

Signed-off-by: John Mulligan <jmulligan@redhat.com>