hack: fix installtool script on MacOS
on MacOS, realpath requires the argument path to exist, and will
fail otherwise. due to this behavior, The installtool script failed on MacOS
(Darwin) and possibly other flavors of Unix and BSD  bacause realpath was
invoked on `$ALT_BIN`, which is essentially the script's first command line
argument,before`ensure_alt_bin` has had a chance to create it.

This change fixes that issue by avoiding the use of realpath before
ensure_alt_bin. This fix is not required on Linux

Signed-off-by: Michael Adam <obnox@samba.org>