github: convert workflows to matrix with OS and arch as dimensions
this expands various existing jobs into multiple jobs by expanding
matrix dimensions for the base operating system (os) ang build
architecture (arch).

It only covers the supported opersting systems Fedora and CentOS
for a start but excludes CentOS for those images that are not supported
yet, i. e. ad-server images. But it can also be considered a preparation
for adding CentOS to the test matrix later when we will be adding
ad-server Containerfiles for centos.

Co-authored-by: John Mulligan <jmulligan@redhat.com>
Signed-off-by: Michael Adam <obnox@samba.org>
Signed-off-by: John Mulligan <jmulligan@redhat.com>
Co-authored-by: Anoop C S <anoopcs@samba.org>
Signed-off-by: Anoop C S <anoopcs@samba.org>