makefile: add build-image that more directly wraps build-image script
The build-*{server,ad-server,client,toolbox} rules are good for humans
and discoverability but are a bit too rigid for how we want to work with
matrix builds in automation. Add a new build-image rule that more
directly wraps the build-image tool and so that it becomes fairly easy
to directly pass the four primary inputs to the tool (kind, package
source, distro base (aka OS), and arch).

Signed-off-by: John Mulligan <jmulligan@redhat.com>