github: use fqin tagging throughout the CI jobs
Use the newer Fully-Qualified-Image-Name tagging scheme in the github ci
actions yaml. This makes the naming of each image better match the core
build parameters (kind, package source, distro base (aka OS), and arch).

The major build and test targets all get an IMG_TAG env var to help
organize and make the rest of the yaml structure consistent.

One workaround we need to apply is to the toolbox builds as the
FROM line in the dockerfiles don't match FQINs nor do they match the
old scheme consistently (see #146). So we just do a little hack
in that section alone to get things working correctly.

Signed-off-by: John Mulligan <jmulligan@redhat.com>