hack: add a --{no-,}distro-qualified option
Add cli options to enable/disable adding "distro qualified" tags to
the images. These are tags like fedora-nightly or centos-latest.

These tags exist mainly to be compatible with the previous versions
of the Makefile before build-image existed and "FQIN"s were being
used. This patch is a first step toward deprecating and then removing
them in favor of FQINs.

Signed-off-by: John Mulligan <jmulligan@redhat.com>