images: do not unset samba_version_suffix on nightly variants
I'm not entirely sure what I was thinking when putting this in
originally - possibly that the var was primarily for doing
release builds and we'd never have nightly builds get released.

However, it's occasionally useful for testing and the var is
unset by default so remove it and simply require that the one
avoid setting samba_version_suffix on nightly builds if one
doesn't want it to apply.

Signed-off-by: John Mulligan <jmulligan@redhat.com>