images: support a package selection distinct from package sources
Previously, you could set a package source such as 'default' or
'nightly' and, along with the distro, it would enable repos and
install a certain packages based on those vars. However, there are times
when having the package selection separate from what repos are being
enabled is handy. This change adds a new variable, that is unused by
default, to directly pick what packages are selected for install.
However, by default, this is still derived from the package source
variable.

Signed-off-by: John Mulligan <jmulligan@redhat.com>