workflows: restructure the push job
Previously, the push job would wait for the builds and tests to finish,
then it would *rebuild* all the images and then push the `latest` and
`nightly` (unqualified) tags to our quay.io container repos.
This change updates the push job so that it re-uses the image builds
that have been built and tested by the previous jobs. It then re-applies
missing tags and updates the repo base. Finally, it pushes both the
unqualified and fully-qualified tags.

Signed-off-by: John Mulligan <jmulligan@redhat.com>