hack: use buildx to cross-build images with docker
When Docker is used and the target architecture differs from the host
architecture, buildx is used for cross-building.

Depending on the container engine and the target architecture,
multiple commands need to be executed to prepare the environment,
build the image, and clean up artifacts afterward.

To streamline this process, a tasks array (an array of lambdas) has
been introduced, its elements are executed sequencally at the end
of the container_build function.

Signed-off-by: Alexander Bachmann <a.t.bachmann@gmail.com>