hack: remove unnecessary uses of lambda
I find lambda to be a bit of an unpleasant thing and try to remove it
when possible. Since all uses of lambda in this code are calling
the same function we can remove the function (and unchanging) argument
and have the variable section map to the run function's remaining
kwargs.

Signed-off-by: John Mulligan <jmulligan@redhat.com>