hack/build-image: add --combined option for building indexes
Indexes (aka manifests) allow for images to be built for different
architectures and the associated with one tag (or set of shared tags).

The build script enables indexes/manifests when supplied with the
--combined flag.

The `build` function is changed a bit and no longer always builds an
image. If the image is already present locally, the build will be
skipped - this is done because (re)building an image of a different
architecture can be very slow. This new behavior tries to avoid costly
unintentional rebuilds. A future change will add a new command to
force a rebuild.

Signed-off-by: John Mulligan <jmulligan@redhat.com>