server: allow passing additional custom repos
Allow passing an arbitrary list of custom repo urls (space separated).
This makes is a lot simpler to create a very custom image.
When using custom repos, it assumes you are taking control of the ceph
deps (as I am) and turns off auto-setup of the ceph repo when
you use --custom-repos.

Signed-off-by: John Mulligan <jmulligan@redhat.com>