hack: support custom container build config files
I constantly rely on my shell history to make custom modifications to
create images based on work-in-progress package builds. Adding a
customizations file allows me to store things in a file so I can run:
```
./hack/build-image \
    -i quay.io/phlogistonjohn/samba-server:custom-centos-amd64-x2 \
    -C foo.ini
```
instead of a long line of embedded junk on the shell. No docs yet
because this is for me. :-)

Signed-off-by: John Mulligan <jmulligan@redhat.com>