hack: add support for sambacc customizing sambacc
Add support for adding a sambacc section to the customize config that
will create a modified version of the container file. This modified
container builds sambacc packages from a git branch (or tag) and then
installs that instead of the default install source.

It is basically a code version of the manual instructions [1] in the
repo.
[1] https://github.com/samba-in-kubernetes/samba-container/blob/master/docs/development.md#build-stage

Signed-off-by: John Mulligan <jmulligan@redhat.com>