images/server: Fix ceph shaman base url parameters
The parameters for the base shaman url that was used to determine the
chacra repository link did not have `status=ready` resulting in invalid
chacra urls as part of the json response. Make sure that we fetch with
`status=ready` so as to create repo files with valid base url.

Signed-off-by: Anoop C S <anoopcs@samba.org>