images/server: Remove explicit installation of dependencies
These packages[1] are automatically pulled in with python3-sambacc
extras which gets built by default and available for installation
as weak dependencies. If we know that these packages[1] exists and
available, then we should be good here to assume that they are pulled
in during the installation of corresponding python3-sambacc extras.

$ rpm -q --requires \
    python3-sambacc+toml-0.8~1.g52a3e0101-1.el9.noarch.rpm | grep tomli
python3.9dist(tomli)

$ rpm -q --requires \
    python3-sambacc+yaml-0.8~1.g52a3e0101-1.el9.noarch.rpm | grep pyyaml
python3.9dist(pyyaml) >= 5.4

$ rpm -q --requires \
    python3-sambacc+rados-0.8~1.g52a3e0101-1.el9.noarch.rpm | grep rados
python3.9dist(rados)

[1] python3-pyyaml, python3-tomli and python3-rados

Signed-off-by: Anoop C S <anoopcs@samba.org>