workflows: clean up ad server k8s test a bit
Take explicit control of parts of the k3d based test.
Begin the process of removing the ishworkh actions which
see little maintenance and are very much a simple wrapper around
actions/{upload,download}-artifact and {docker,podman} {save,load}
commands. This way we being the process of using maintained and
pinned github actions dependencies.

NOTE: Originally I was also going to bypass the docker load
and just load the tarballs directly to k3d but that seems to
confuse it WRT tags. In the future, someone could look into
bypassing docker entirely to make this a little more efficient.

Signed-off-by: John Mulligan <jmulligan@redhat.com>