workflows: pin the k3d setup action version
I would have preferred to replace this action as well but unlike the
ishworkh actions what it's doing is not as simple and I am not well
versed in k3d. For the short term we can just pin the version and
perhaps in the future we can replace it or replace k3d with something we
(I) actually understand and don't dread having to debug.

Signed-off-by: John Mulligan <jmulligan@redhat.com>