github: update mergify rules to match new checks
Update the mergify rules to match the new checks based on the
recently updated build matrix.

This list is generated by a short script I wrote to pull the checks from
the github api (using PyGithub library) and then tweaked the yaml to
match the kind of lines needed by the mergify config.

Signed-off-by: John Mulligan <jmulligan@redhat.com>