github: add various new build condition checks to mergify rules
With the addition of various arm64 jobs as well as some other jobs that
we had added in the past, update the mergify rules to match the new
checks based on the recently updated build matrix.

The list of jobs were was generated by a short script I wrote to query
the checks from the github api (using PyGithub library) and then tweaked
the yaml to match the kind of lines needed by the mergify config.
Because the previous two commits organized and sorted the lists it's
now easy to see what jobs have been added now.

Signed-off-by: John Mulligan <jmulligan@redhat.com>